refactor(context-menu): isolate IPC and UI
