+++ src/renderer/app/directives/context-menu-ipc/context-menu-ipc.less
@@ -1,0 +1,3 @@
+context-menu-ipc {
+    display: none;
+}
