--- src/renderer/app/directives/context-menu/context-menu.controller.ts
@@ -1,1 +1,0 @@
-export class ContextMenuController {}
