+++ src/renderer/app/directives/context-menu/context-menu.directive.ts
@@ -1,137 +1,96 @@
-import { IDirective, IDirectiveFactory, IRootScopeService, IScope } from 'angular'
+import { IDirective, IDirectiveFactory, IScope } from 'angular'
 
-import type { ColorTheme, ContextMenuItemModel } from '@shared/ipc-contract'
+import type { ContextMenuItemModel, ContextMenuModel, ContextMenuPlacement, ContextMenuSize, Unsubscribe } from '@shared/ipc-contract'
+import template from './context-menu.template.html'
 
-interface ContextMenuItem {
-    label: string
-    icon?: string
-    role?: string
-    click?: (...args: any[]) => void
-    check?: (item: any) => boolean
-    menu?: ContextMenuItem[]
+interface ContextMenuWindowBridge {
+    onModel(callback: (model: ContextMenuModel) => void): Unsubscribe
+    resize(size: ContextMenuSize): Promise<ContextMenuPlacement>
+    hide(): Promise<void>
+    select(actionId: string): Promise<void>
 }
+
+declare const window: Window & { contextMenu: ContextMenuWindowBridge }
 
 interface ContextMenuScope extends IScope {
-    menu: ContextMenuItem[]
-    selectedItems: any[]
-    bind: {
-        show: (event: MouseEvent, items: any[]) => void
-        hide: () => void
-    }
-    click: (...args: any[]) => void
-    debug?: () => void
+    menu: ContextMenuItemModel[]
+    debugItem?: ContextMenuItemModel
+    runMenuItem(item: ContextMenuItemModel): void
+    toggleSubmenu(event: Event, visible: boolean): void
 }
+
+const MENU_WINDOW_MARGIN = 12
 
 export class ContextMenuDirective implements IDirective {
     restrict = 'E'
-    scope = {
-        menu: '=',
-        bind: '=',
-        click: '=',
-        debug: '=?',
-    }
+    scope = {}
+    template = template
 
     static getInstance(): IDirectiveFactory {
-        const factory = ($rootScope: IRootScopeService) => new ContextMenuDirective($rootScope)
-        factory.$inject = ['$rootScope']
+        const factory = ($timeout: ng.ITimeoutService) => new ContextMenuDirective($timeout)
+        factory.$inject = ['$timeout']
         return factory
     }
 
-    constructor(private $rootScope: IRootScopeService) {}
+    constructor(private $timeout: ng.ITimeoutService) {}
 
-    link(scope: ContextMenuScope) {
-        let actions = new Map<string, ContextMenuItem>()
-        let nextActionId = 0
-
-        const isItemVisible = (item: ContextMenuItem) => {
-            if (item.menu) return true
-
-            const features = this.$rootScope.$btclient?.features
-            switch (item.role) {
-                case 'details': return !!features?.torrentDetails
-                case 'set-speed-limits': return !!features?.speedLimits
-                case 'set-ratio': return !!features?.ratioLimits
-                default: return true
+    link(scope: ContextMenuScope, element: ng.IAugmentedJQuery) {
+        scope.menu = []
+        scope.runMenuItem = (item) => {
+            if (item.id) {
+                void window.contextMenu.select(item.id)
             }
         }
-
-        const serializeItem = (item: ContextMenuItem): ContextMenuItemModel | null => {
-            if (!isItemVisible(item)) return null
-
-            if (item.menu) {
-                return {
-                    label: item.label,
-                    icon: item.icon,
-                    role: item.role,
-                    menu: item.menu.map(serializeItem).filter((child): child is ContextMenuItemModel => !!child),
-                }
-            }
-
-            const id = `context-action-${nextActionId++}`
-            actions.set(id, item)
-            return {
-                id,
-                label: item.label,
-                icon: item.icon,
-                role: item.role,
-                checked: item.check ? scope.selectedItems.every((entry) => item.check!(entry)) : undefined,
-            }
+        scope.toggleSubmenu = (event, visible) => {
+            angular.element(event.currentTarget).find('.menu').css('display', visible ? 'block' : 'none')
         }
 
-        const hide = () => {
-            void window.electorrent.contextMenu.hide()
+        const closeWhenOutsideMenu = (event: MouseEvent) => {
+            if (!(event.target as Element).closest('.ui.menu')) {
+                void window.contextMenu.hide()
+            }
         }
+        document.addEventListener('click', closeWhenOutsideMenu)
 
-        scope.bind = {
-            show: (event: MouseEvent, items: any[]) => {
-                scope.selectedItems = items
-                actions = new Map()
-                nextActionId = 0
-                const menu = scope.menu.map(serializeItem).filter((item): item is ContextMenuItemModel => !!item)
-                const debugItem = scope.debug && this.$rootScope.$btclient
-                    ? serializeItem({ label: 'Debug', icon: 'help', role: 'debug', click: scope.debug }) || undefined
-                    : undefined
-                const themeLink = document.querySelector<HTMLLinkElement>('link[href*="css/themes/"]')
-                const theme: ColorTheme = themeLink?.href.includes('/dark.css') ? 'dark' : 'light'
+        const unsubscribe = window.contextMenu.onModel((model) => {
+            const stylesheet = document.createElement('link')
+            stylesheet.rel = 'stylesheet'
+            stylesheet.href = `css/themes/${model.theme}.css`
+            document.head.appendChild(stylesheet)
 
-                void window.electorrent.app.getMeta().then((meta) => {
-                    void window.electorrent.contextMenu.show({
-                        x: event.clientX,
-                        y: event.clientY,
-                        theme,
-                        items: menu,
-                        debugItem: meta.isDebug ? debugItem : undefined,
+            stylesheet.addEventListener('load', () => {
+                scope.$applyAsync(() => {
+                    scope.menu = model.items
+                    scope.debugItem = model.debugItem
+                    this.$timeout(() => {
+                        const root = element[0]
+                        const primaryMenu = root.querySelector<HTMLElement>(':scope > .ui.menu')
+                        const submenus = Array.from(root.querySelectorAll<HTMLElement>('.context.dropdown > .menu'))
+                        const submenuSize = submenus.reduce((size, submenu) => {
+                            const display = submenu.style.display
+                            const visibility = submenu.style.visibility
+                            submenu.style.visibility = 'hidden'
+                            submenu.style.display = 'block'
+                            size.width = Math.max(size.width, submenu.scrollWidth)
+                            size.height = Math.max(size.height, submenu.scrollHeight)
+                            submenu.style.display = display
+                            submenu.style.visibility = visibility
+                            return size
+                        }, { width: 0, height: 0 })
+                        void window.contextMenu.resize({
+                            width: Math.ceil((primaryMenu?.scrollWidth || root.scrollWidth) + submenuSize.width + (MENU_WINDOW_MARGIN * 2) + 2),
+                            height: Math.ceil(Math.max(primaryMenu?.scrollHeight || root.scrollHeight, submenuSize.height) + (MENU_WINDOW_MARGIN * 2) + 2),
+                        }).then(({ submenuOnLeft }) => {
+                            element.toggleClass('submenu-on-left', submenuOnLeft)
+                        })
                     })
                 })
-            },
-            hide,
-        }
-
-        const unsubscribe = window.electorrent.contextMenu.onAction((actionId) => {
-            const item = actions.get(actionId)
-            if (item) {
-                scope.$applyAsync(() => scope.click(item.click, item.label, item))
-            }
+            }, { once: true })
         })
-        const onPointerDown = () => hide()
-        const onKeyUp = (event: KeyboardEvent) => {
-            if (event.key === 'Escape') hide()
-        }
-        const onScroll = () => hide()
-        const onResize = () => hide()
-
-        document.body.addEventListener('pointerdown', onPointerDown, true)
-        document.addEventListener('keyup', onKeyUp)
-        document.querySelector('.main-content')?.addEventListener('scroll', onScroll)
-        window.addEventListener('resize', onResize)
 
         scope.$on('$destroy', () => {
             unsubscribe()
-            document.body.removeEventListener('pointerdown', onPointerDown, true)
-            document.removeEventListener('keyup', onKeyUp)
-            document.querySelector('.main-content')?.removeEventListener('scroll', onScroll)
-            window.removeEventListener('resize', onResize)
-            hide()
+            document.removeEventListener('click', closeWhenOutsideMenu)
         })
     }
 }
