+++ src/renderer/app/directives/context-menu/context-menu.less
@@ -1,8 +1,4 @@
 context-menu {
-    position: fixed;
-    z-index: 300;
-    display: none;
-
     .ui.vertical.menu {
         border: 0;
         box-shadow: 0 3px 8px rgba(0, 0, 0, .3);
