+++ src/renderer/app/directives/torrents-page/torrents-page.template.html
@@ -150,8 +150,8 @@
         </div>
     </div>
 
-    <context-menu id="contextmenu" bind="contextMenu" menu="$btclient.contextMenu" click="doContextAction" debug="debug">
-    </context-menu>
+    <context-menu-ipc bind="contextMenu" menu="$btclient.contextMenu" click="doContextAction" debug="debug">
+    </context-menu-ipc>
 
     <set-label-modal
         labels="labels"
