+++ test/e2e/e2e_torrent.ts
@@ -365,11 +365,12 @@ export class Torrent {
 
     const parentWindow = await browser.getWindowHandle()
     for (let attempt = 0; attempt < 3; attempt++) {
+      const existingWindowHandles = new Set(await browser.getWindowHandles())
       await elem.click(options)
       try {
         const contextMenuWindow = await browser.waitUntil(async () => {
           const handles = await browser.getWindowHandles()
-          return handles.find((handle) => handle !== parentWindow)
+          return handles.find((handle) => !existingWindowHandles.has(handle))
         }, { timeout: attempt === 2 ? this.timeout : 1000 })
         await browser.switchToWindow(contextMenuWindow)
         const contextMenu = $("#contextmenu > .ui.menu")
